% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background}
\alias{background}
\title{A background color to be used for pages. It is off white}
\format{
An object of class \code{character} of length 1.
}
\usage{
background
}
\description{
A background color to be used for pages. It is off white
}
\keyword{datasets}
