% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{component_example}
\alias{component_example}
\title{component_example
re exported from shiny.mui}
\format{
An object of class \code{function} of length 1.
}
\usage{
component_example
}
\description{
component_example
re exported from shiny.mui
}
\keyword{datasets}
