% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QGrid.R
\name{Container}
\alias{Container}
\title{Helper to create grid containers}
\usage{
Container(...)
}
\description{
Helper to create grid containers
}
\examples{
library(shiny.quartz)
library(shiny.mui)

component_example(
    Container(Item(xs = 6, "hello"), Item(xs = 6, "world"))
)
}
