% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{format_ci}
\alias{format_ci}
\title{format_ci
Pretty print CI}
\usage{
format_ci(val, lower, upper)
}
\arguments{
\item{val}{main value}

\item{lower}{lower CI}

\item{upper}{upper CI}
}
\description{
format_ci
Pretty print CI
}
\examples{
format_ci(0.5, 0.1, 0.9)
}
