% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HStack.R
\name{HStack}
\alias{HStack}
\title{Horizontal Stack}
\usage{
HStack(..., sx = list())
}
\arguments{
\item{...}{A list of components to stack.}

\item{sx}{A list of style attributes to apply to the component.}
}
\description{
A layout component for stacking horizontally with gaps.
}
\examples{
component_example(HStack(tags$button("hi", "HI"), tags$button("hi", "hi")))
}
\seealso{
Other layout: 
\code{\link{Container}()},
\code{\link{Item}()},
\code{\link{Page}()},
\code{\link{QCard}()},
\code{\link{VStack}()}
}
\concept{layout}
