% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{IconSwitch.shinyInput}
\alias{IconSwitch.shinyInput}
\title{Icon Switch}
\usage{
IconSwitch.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{inputId}{The input id}

\item{...}{args to pass to the element including the icon for the switch}

\item{value}{The value of the input}
}
\description{
Adds a switch for boolean values
}
\examples{
library(shiny.quartz)
library(shiny.mui)
library(shiny)


if (interactive()) {
    shinyApp(
        ui = div(
            IconSwitch.shinyInput("text", TableViewIcon(), value = T),
            textOutput("textValue")
        ),
        server = function(input, output) {
            output$textValue <- renderText({
                sprintf("Value: \%s", input$text)
            })
        }
    )
}
}
