% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoPopover.R
\name{InfoPopover}
\alias{InfoPopover}
\title{Info Popover}
\usage{
InfoPopover(...)
}
\arguments{
\item{...}{Children to include in the popover.}
}
\description{
Provides a icon that shows a popover with children when clicked.
}
\examples{
component_example(InfoPopover("Hi", "Hello"))
}
