% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{input}
\alias{input}
\title{Create Shiny Input Component from React JS}
\usage{
input(name, defaultValue = "")
}
\arguments{
\item{name}{The name of the component.}

\item{defaultValue}{the default value of the component.}
}
\description{
Create Shiny Input Component from React JS
}
