% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QGrid.R
\name{Item}
\alias{Item}
\title{Item}
\usage{
Item(...)
}
\arguments{
\item{...}{other props to pass to the react component.}
}
\description{
Helper to create grid items. Width is 12 by default.
}
\examples{
library(shiny.quartz)
library(shiny.mui)

component_example(
    Container(Item(xs = 6, "hello"), Item(xs = 6, "world"))
)
}
\seealso{
Other layout: 
\code{\link{Container}()},
\code{\link{HStack}()},
\code{\link{Page}()},
\code{\link{QCard}()},
\code{\link{VStack}()}

Other grid: 
\code{\link{Container}()}
}
\concept{grid}
\concept{layout}
