% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_options.R
\name{make_options}
\alias{make_options}
\title{Make options for QSelect}
\usage{
make_options(...)
}
\arguments{
\item{...}{A vector of strings.}
}
\description{
Converts vectors to a list of keyword pairs.
}
\examples{
make_options(1:5)
}
\seealso{
\code{\link[=QSelect.shinyInput]{QSelect.shinyInput()}}
}
