% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page.R
\name{Page}
\alias{Page}
\title{Page}
\usage{
Page(title, ...)
}
\arguments{
\item{title}{character the title of the page.}

\item{...}{a list of components to include in the page.}
}
\description{
Generates a page with a gray background and a title at the top.
Also applies qthemeprovider to the page.
}
\examples{
component_example(Page("Title", "Hello there"))

}
\seealso{
\code{\link[=QThemeProvider]{QThemeProvider()}}

Other layout: 
\code{\link{Container}()},
\code{\link{HStack}()},
\code{\link{Item}()},
\code{\link{QCard}()},
\code{\link{VStack}()}
}
\concept{layout}
