% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette}
\alias{palette}
\title{A quartz palette to be used for plots}
\format{
An object of class \code{character} of length 4.
}
\usage{
palette
}
\description{
A quartz palette to be used for plots
}
\keyword{datasets}
