% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAccordion.R
\name{QAccordion}
\alias{QAccordion}
\title{Quartz Accordion}
\usage{
QAccordion(title, ...)
}
\arguments{
\item{title}{The title of the accordion}

\item{...}{The inner content of the accordion}
}
\description{
A simplified version of the default accordion built with MUI
}
\examples{
component_example(QAccordion(title = shiny.mui::Typography("Title"), shiny.mui::Typography("Hello")))
}
