% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCard.R
\name{QCard}
\alias{QCard}
\title{Helper to create cards with card content}
\usage{
QCard(..., title = "", Toolbar = Box())
}
\arguments{
\item{...}{Children to include in the card}

\item{title}{Title of the card}

\item{Toolbar}{optional toolbar to include in the card}
}
\description{
Helper to create cards with card content
}
\examples{
component_example(QCard("Hello there", title = "Card"))
}
\seealso{
\code{\link[shiny.mui:Card]{shiny.mui::Card()}}

Other layout: 
\code{\link{Container}()},
\code{\link{HStack}()},
\code{\link{Item}()},
\code{\link{Page}()},
\code{\link{VStack}()}
}
\concept{layout}
