% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDataGrid.R
\name{QDataGrid}
\alias{QDataGrid}
\title{QDataGrid}
\usage{
QDataGrid(data, height = "400px", columns = NULL, minColWidth = 100, ...)
}
\arguments{
\item{data}{data_frame to display}

\item{height}{height of the grid}

\item{columns}{configuration for individual columns. See the mui documentation for more details \href{https://mui.com/x/react-data-grid/column-definition/}{here}}

\item{minColWidth}{minimum width of columns}

\item{...}{other props to pass to the react component. See \href{https://mui.com/x/react-data-grid/}{mui documentation}}
}
\description{
Helper to create data grids
}
\examples{
library(shiny.mui)
library(shiny)
library(shiny.quartz)
data(mtcars)
library(dplyr)

if (interactive()) {
    shinyApp(
        ui = div(
            reactOutput("data")
        ),
        server = function(input, output) {
            output$data <- renderReact({
                QDataGrid(mtcars, minColWidth = 100)
            })
        }
    )
}
}
