% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QListItem.R
\name{QListItem}
\alias{QListItem}
\title{Q List Item}
\usage{
QListItem(primary, ..., secondary = "")
}
\arguments{
\item{primary}{primary text}

\item{...}{argumens to pass to shiny.quartz ListItem}

\item{secondary}{secondary text}
}
\description{
Helper to create shiny.mui list items with primary and secondary text
}
\examples{
library(shiny.quartz)
library(shiny.mui)

component_example(
    List(QListItem("first", secondary = "go"), QListItem("second"), QListItem("third"))
)
}
\seealso{
\code{\link[shiny.mui:List]{shiny.mui::List()}}
}
