% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSelect.R
\name{QSelect.shinyInput}
\alias{QSelect.shinyInput}
\title{Q Select}
\usage{
QSelect.shinyInput(..., options, label = "")
}
\arguments{
\item{...}{args to pass to the element shiny.mui::Select.shinyInput() element}

\item{options}{A list of key value pairs}

\item{label}{label to display for the select}
}
\description{
Warning. This will not render correctly if not attatched to a shiny server.
}
\examples{
library(shiny.quartz)
library(shiny)


if (interactive()) {
    shinyApp(
        ui = div(
            QSelect.shinyInput("text", fullWidth = T, label = "First One", value = "two", options = make_options("one", "two", "three")),
            textOutput("textValue")
        ),
        server = function(input, output) {
            output$textValue <- renderText({
                sprintf("Value: \%s", input$text)
            })
        }
    )
}
}
