% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSwitch.R
\name{QSwitch.shinyInput}
\alias{QSwitch.shinyInput}
\title{QSwitch}
\usage{
QSwitch.shinyInput(..., label)
}
\arguments{
\item{...}{other props to pass to the react component.}

\item{label}{label to display}
}
\description{
Helper to add labels to shiny.mui switches
}
\seealso{
\code{\link[shiny.mui:Switch]{shiny.mui::Switch()}}
}
