% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QThemeProvider.R
\name{QThemeProvider}
\alias{QThemeProvider}
\title{Provider for Quartz Software MUI Theme}
\usage{
QThemeProvider(...)
}
\arguments{
\item{...}{aditional args to pass to the ThemeProvider}
}
\description{
Provider for Quartz Software MUI Theme
}
\examples{
component_example(QThemeProvider(shiny.mui::Button("Hello World")))
}
