% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggleButton.R
\name{renderToggleButton}
\alias{renderToggleButton}
\title{Toggleable button}
\usage{
renderToggleButton(Icon, open, input)
}
\arguments{
\item{Icon}{the icon for the button}

\item{open}{a reactive val}
}
\description{
Toggleable button
}
