% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleSelect.R
\name{SimpleSelect.shinyInput}
\alias{SimpleSelect.shinyInput}
\title{Simple Select}
\usage{
SimpleSelect.shinyInput(..., options)
}
\arguments{
\item{options}{A list of key value pairs}
}
\description{
Warning. This will not render correctly if not attatched to a shiny server.
}
