% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_app}
\alias{use_app}
\title{Uses the default app for quartz packages}
\usage{
use_app()
}
\description{
This is needed for hosting on shinyapps.io
}
\seealso{
Other utils: 
\code{\link{use_dev}()},
\code{\link{use_mod}()},
\code{\link{use_run_app}()}
}
\concept{utils}
