% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_dev}
\alias{use_dev}
\title{Add the dev file to run development version of app}
\usage{
use_dev()
}
\description{
Add the dev file to run development version of app
}
\seealso{
Other utils: 
\code{\link{use_app}()},
\code{\link{use_mod}()},
\code{\link{use_run_app}()}
}
\concept{utils}
