% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_run_app}
\alias{use_run_app}
\title{Uses the default runapp for quartz packages}
\usage{
use_run_app()
}
\description{
Uses the default runapp for quartz packages
}
\seealso{
Other utils: 
\code{\link{use_app}()},
\code{\link{use_dev}()},
\code{\link{use_mod}()}
}
\concept{utils}
