% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VStack.R
\name{VStack}
\alias{VStack}
\title{vstack}
\usage{
VStack(..., sx = list())
}
\description{
A layout component for stacking vertically with gaps.
}
\examples{
library(shiny.quartz)
library(shiny.mui)

Cube <- shiny.mui::Box(height = 100, width = "100\%", backgroundColor = "red")
component_example(
    VStack(sx = list(gap = 2), Cube, Cube)
)
}
