% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcknowledgementCard.R
\name{AcknowledgementCard}
\alias{AcknowledgementCard}
\title{A card for acknowledging the authors of a publication.}
\usage{
AcknowledgementCard(
  authors = list(list(name = "Timothy L. Wiemken", src =
    "https://media-exp1.licdn.com/dms/image/C4E03AQE4JkLxHi1g7A/profile-displayphoto-shrink_400_400/0/1603549767246?e=1655337600&v=beta&t=mGGA7RxgQvdi0yIvbwiOJXiJ0uvM_YAm0-UuybyOp8g"),
    list(name = "Jacob A. Clarke", src =
    "https://www.jacobaclarke.com/headshot_new_cropped.jpg", href =
    "https://www.jacobaclarke.com/"), list(name = "Christopher Prener"), list(name =
    "Verna Welch"), list(name = "John McLaughlin"))
)
}
\arguments{
\item{authors}{A list of author names. List can contain name, src, and href.}
}
\description{
A card for acknowledging the authors of a publication.
}
\examples{
component_example(AcknowledgementCard())
}
