% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcknowledgementCard.R
\name{AcknowledgementCard}
\alias{AcknowledgementCard}
\title{A card for acknowledging the authors of a publication.}
\usage{
AcknowledgementCard(
  authors = c("Timothy L. Wiemken", "Jacob A. Clarke", "Christopher Prener",
    "John McLaughlin")
)
}
\arguments{
\item{authors}{A vector of author names.}
}
\description{
A card for acknowledging the authors of a publication.
}
\examples{
library(shiny.quartz)
library(shiny.mui)

component_example(
    AcknowledgementCard()
)
}
