% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page.R
\name{Page}
\alias{Page}
\title{Page}
\usage{
Page(title, ...)
}
\description{
Page
}
\examples{
library(shiny.quartz)
library(shiny.mui)

component_example(
    Page(
        "Title",
        TextField.shinyInput("test", label = "Should be Full Width")
    )
)
}
