% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCard.R
\name{QCard}
\alias{QCard}
\title{Helper to create cards with card content}
\usage{
QCard(..., title = "")
}
\description{
Helper to create cards with card content
}
\examples{
library(shiny.quartz)
library(shiny.mui)

component_example(
    QCard(title = "Hello", "go")
)
}
