% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDataGrid.R
\name{QDataGrid}
\alias{QDataGrid}
\title{Helper to create data grids}
\usage{
QDataGrid(data, height = "400px", ...)
}
\description{
Helper to create data grids
}
\examples{
library(shiny.mui)
library(shiny)
library(shiny.quartz)
data(mtcars)


if (interactive()) {
    shinyApp(
        ui = div(
            reactOutput("data")
        ),
        server = function(input, output) {
            output$data <- renderReact({
                QDataGrid(mtcars)
            })
        }
    )
}
}
