% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QListItem.R
\name{QListItem}
\alias{QListItem}
\title{Q List}
\usage{
QListItem(primary, ..., secondary = "")
}
\description{
Q List
}
\examples{
library(shiny.quartz)
library(shiny.mui)

component_example(
    List(QListItem("first"), QListItem("second"), QListItem("third"))
)
}
