% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ReactMarkdown}
\alias{ReactMarkdown}
\title{ReactMarkdown}
\usage{
ReactMarkdown(...)
}
\arguments{
\item{...}{args to pass to react markdown}
}
\description{
Render inner markdown text. This comes from the \code{react-markdown} package. https://github.com/remarkjs/react-markdown
}
