% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VStack.R
\name{VStack}
\alias{VStack}
\title{vstack}
\usage{
VStack(..., sx = list())
}
\description{
A layout component for stacking vertically with gaps.
}
\examples{
component_example(VStack(tags$button("hi", "HI"), tags$button("hi", "hi")))
}
\seealso{
Other layout: 
\code{\link{Container}()},
\code{\link{HStack}()},
\code{\link{Item}()},
\code{\link{Page}()},
\code{\link{QCard}()}
}
\concept{layout}
